Sprint `component-context.ts`
