+++ src/component/component-context.key.ts
@@ -1,0 +1,7 @@
+import { ContextKey, SingleContextKey } from 'context-values';
+import { ComponentContext } from './component-context';
+
+/**
+ * @internal
+ */
+export const componentContextKey: ContextKey<ComponentContext<any>> = new SingleContextKey('component-context');
