+++ src/component/component-event.ts
@@ -1,4 +1,6 @@
+import { ContextKey } from 'context-values';
 import { ComponentContext } from './component-context';
+import { componentEventDispatcherKey } from './component-event.key';
 
 /**
  * Component event.
@@ -30,5 +32,24 @@ export class ComponentEvent extends Event {
   constructor(type: string, eventInitDict?: EventInit) {
     super(type, eventInitDict);
   }
+
+}
+
+/**
+ * Component event dispatcher function is used to dispatch component events.
+ *
+ * It is available in bootstrap context context.
+ *
+ * @param context A context of component to dispatch an `event` for.
+ * @param event An event to dispatch.
+ */
+export type ComponentEventDispatcher = (context: ComponentContext<any>, event: Event) => void;
+
+export namespace ComponentEventDispatcher {
+
+  /**
+   * A key of bootstrap context value containing component event dispatcher.
+   */
+  export const key: ContextKey<ComponentEventDispatcher> = componentEventDispatcherKey;
 
 }
