+++ src/component/index.ts
@@ -4,3 +4,4 @@ export * from './component-context';
 export * from './component-def';
 export * from './component-event';
 export * from './component-mount';
+export * from './content-root';
